Refactoring of learningObject and userServices
