+++ src/app/cube/organization-list/user-card/user-card.component.html
@@ -1,6 +1,6 @@
 <div class="user-card" [routerLink]="['/users', user._username]">
   <div class="image">
-    <img [src]="getGravatarImage()" class="img">
+    <img [src]="icon" class="img">
   </div>
   <div class="user-content">
     <div class="title">{{user.name}}</div>
