+++ src/app/cube/user-profile/user-profile.component.ts
@@ -1,3 +1,4 @@
+import { UserService } from './../../core/user.service';
 import { USER_ROUTES } from '../../../environments/route';
 import { Component, OnInit, OnDestroy} from '@angular/core';
 import { LearningObjectService } from '../learning-object.service';
@@ -25,11 +26,12 @@ export class UserProfileComponent implements OnInit {
   height: number = 100;
   width: number = 100;
   size: number = 200; 
-  default: string; 
+  gravatarImage: string; 
   editContent: boolean = false;
 
   constructor(
-    private service: LearningObjectService,
+    private learningObjectService: LearningObjectService,
+    private userService: UserService,
     private auth: AuthService,
     private modalService: ModalService,
     private route: ActivatedRoute,
@@ -40,8 +42,9 @@ export class UserProfileComponent implements OnInit {
     this.subscription = this.route.data.subscribe(val => {
       this.user = val.user;
       this.self = this.user.username === this.auth.username;
-      this.default = 'identicon';
     });
+    
+    this.gravatarImage = this.userService.getGravatarImage(this.user.email,this.size);
 
     // particle config
     this.myStyle = {
@@ -176,11 +179,5 @@ export class UserProfileComponent implements OnInit {
     if (changed) {
       this.user = this.auth.user;
     }
-  }
-
-  getGravatarImage():string {
-    // r=pg checks the rating of the Gravatar image 
-    return 'http://www.gravatar.com/avatar/' + md5(this.user.email) + '?s=' + this.size + 
-      '?r=pg&d=' + this.default;
   }
 }
